Still addressing issue #5.
Added library to store loaded schemas. Instead of loading the same schemas repeatedly, every time a descriptor needs to be validated, schemas
are stored in local memory to increase efficiency. The function 'load_schema()' is responsible for retrieving schemas, which may be local or
remote. If the required schema was previously loaded it will return it from the memory cache. To force a schema reload, the reload parameter
must be set to True.

The validators library was added to the buildout configuration as a project dependency (https://github.com/kvesteri/validators/). Validators is used to validate the provided URL
to load schemas.